refactor: Move background fetching to server
- move background related functions to its module
- remove irrelevant settings
